Align Integer/Number fidelity for sum, mod, round and trunc.
Keep aggregate sum on Integer as Long, force analytic window sum to Number, and stop Spark SQL round from overriding VTL Integer return types. TCK Spark3 moves 123→127 passes.

+++ vtl-engine/src/test/java/fr/insee/vtl/engine/processors/InMemoryProcessingEngineAggrTest.java
@@ -46,7 +46,7 @@ class InMemoryProcessingEngineAggrTest {
 
     assertThat(mechanical.getDataStructure().get("me_1").getRole()).isEqualTo(Dataset.Role.MEASURE);
     assertThat(result.getDataStructure()).isEqualTo(plan.structure());
-    assertThat(result.getDataStructure().get("me_1").getType()).isEqualTo(Double.class);
+    assertThat(result.getDataStructure().get("me_1").getType()).isEqualTo(Long.class);
   }
 
   @Test
